+++ mayan/apps/motd/apps.py
@@ -3,4 +3,5 @@ from mayan.apps.common.apps import MayanAppConfig
 
 class MOTDApp(MayanAppConfig):
     has_tests = False
+    has_translations = False
     name = 'mayan.apps.motd'